% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{binary_01}
\alias{binary_01}
\title{Binary recode reversed (1=0, 2=1)}
\format{
A named numeric vector of length 2.
}
\usage{
binary_01
}
\description{
Named numeric vector for recoding 1/2 survey responses to 0/1.
}
\keyword{datasets}
