% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{binary_10}
\alias{binary_10}
\title{Binary recode (1=1, 2=0)}
\format{
A named numeric vector of length 2.
}
\usage{
binary_10
}
\description{
Named numeric vector for recoding 1/2 survey responses to 1/0.
}
\keyword{datasets}
