% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{five_n}
\alias{five_n}
\title{Normal-coded 5-point Likert scale}
\format{
A named numeric vector of length 5.
}
\usage{
five_n
}
\description{
Named numeric vector preserving a 1-5 scale (identity mapping).
}
\keyword{datasets}
