% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{four_n}
\alias{four_n}
\title{Normal-coded 4-point Likert scale}
\format{
A named numeric vector of length 4.
}
\usage{
four_n
}
\description{
Named numeric vector preserving a 1-4 scale (identity mapping).
}
\keyword{datasets}
