% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{four_r}
\alias{four_r}
\title{Reverse-coded 4-point Likert scale}
\format{
A named numeric vector of length 4.
}
\usage{
four_r
}
\description{
Named numeric vector that reverses a 1-4 scale (1→4, 2→3, 3→2, 4→1).
}
\keyword{datasets}
