% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{infer_metric}
\alias{infer_metric}
\title{Infer metric label from a recode_rules vector}
\usage{
infer_metric(rules)
}
\arguments{
\item{rules}{A named character/numeric vector of recode rules.}
}
\value{
A character string describing the metric.
}
\description{
Infer metric label from a recode_rules vector
}
\keyword{internal}
