% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{render_codebook}
\alias{render_codebook}
\title{Render a codebook data frame as a gt table}
\usage{
render_codebook(cb, title = "Variable Codebook", has_stats = FALSE)
}
\arguments{
\item{cb}{A codebook tibble from \code{\link[=codebook]{codebook()}}.}

\item{title}{Table title.}

\item{has_stats}{Logical; whether summary stats columns are present.}
}
\value{
A \code{gt} object.
}
\description{
Render a codebook data frame as a gt table
}
\keyword{internal}
