% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_brms.R
\name{run_brms_app}
\alias{run_brms_app}
\title{Launch BRMS Regression Shiny App}
\usage{
run_brms_app(sample_data = FALSE, ...)
}
\arguments{
\item{sample_data}{Optional. If TRUE, loads mtcars into global environment}

\item{...}{Additional arguments passed to shiny::runApp}
}
\description{
Launch BRMS Regression Shiny App
}
\examples{
if (interactive()) {
  # Load some data first
  data(mtcars)
  data(iris)

  # Then run the app
  run_brms_app()
}
}
