% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{seven_n}
\alias{seven_n}
\title{Normal-coded 7-point scale}
\format{
A named numeric vector of length 7.
}
\usage{
seven_n
}
\description{
Named numeric vector preserving a 1-7 scale (identity mapping).
}
\keyword{datasets}
