% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_helpers.R
\name{summarize_missingness}
\alias{summarize_missingness}
\title{Summarize missingness in a data frame}
\usage{
summarize_missingness(data, vars = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{vars}{Character vector of column names. If NULL, all columns are used.}
}
\value{
A data.frame with columns Variable, N, Missing, Pct_Missing
}
\description{
Summarize missingness in a data frame
}
