% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_common_recodes.R
\docType{data}
\name{three_n}
\alias{three_n}
\title{Normal-coded 3-point scale}
\format{
A named numeric vector of length 3.
}
\usage{
three_n
}
\description{
Named numeric vector preserving a 1-3 scale (identity mapping).
}
\keyword{datasets}
