% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmOrdinal.R
\name{brms.ordinal}
\alias{brms.ordinal}
\title{brms.ordinal}
\usage{
brms.ordinal(
  data = dataActive,
  IV = IV,
  DV = DV,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  cores = 10,
  seed = 1234,
  ...
)
}
\arguments{
\item{Specify}{DV, IV, data, and other arguments native to BRMS.
For example, with +1 IV and/or two way interactions;
IV <- "presvote_trump_2020 + prepost + presvote_trump_2020:prepost"}
}
\value{
brms model object
}
\description{
Estimate an ordinal regression model in \code{brms}
}
